fix(Native): Android assetsBundleFileName null (RN 0.74.7 Fabric Bridgeless release)
I found an issue in the Android RN 0.74.7 (New Architecture + Bridgeless) release build where the CodePush instance is newly created when reloading or updating downloads, causing the JS bundle file name to be reset to null.

Since the CodePush instance is not newly created in Old Architecture and New Architecture Bridge mode, I modified it to be a singleton in Bridgeless mode as well.